Total Travel Costs are now ints
The distance matrix is a uniform distribution of [0,1000] making floats unecessary.

+++ src/experimental_results.jl
@@ -22,7 +22,7 @@ results = DataFrame(
     "Improved? H2" => String[],
 )
 
-function experiments_table(iteration, total_travel_cost_h1::Float64, recollected_prize_h1::Int, improved_travel_cost_h1::Float64, optimal_value_h1::Float64, optimality_gap_h1::Float64, total_travel_cost_h2::Float64, recollected_prize_h2::Int, improved_travel_cost_h2::Float64, optimal_value_h2::Float64, optimality_gap_h2::Float64)
+function experiments_table(iteration, total_travel_cost_h1::Int, recollected_prize_h1::Int, improved_travel_cost_h1::Int, optimal_value_h1::Float64, optimality_gap_h1::Float64, total_travel_cost_h2::Int, recollected_prize_h2::Int, improved_travel_cost_h2::Int, optimal_value_h2::Float64, optimality_gap_h2::Float64)
     push!(results, (iteration, total_travel_cost_h1, recollected_prize_h1, recollected_prize_h1 / total_travel_cost_h1, improved_travel_cost_h1, optimal_value_h1, optimality_gap_h1, total_travel_cost_h1 > improved_travel_cost_h1 ? "Yes" : "No", total_travel_cost_h2, recollected_prize_h2, recollected_prize_h2 / total_travel_cost_h2, improved_travel_cost_h2, optimal_value_h2, optimality_gap_h2, total_travel_cost_h2 > improved_travel_cost_h2 ? "Yes" : "No"))
     return results
 end
